Tweak things to prevent spec from hanging when it fails
The only thing here that actually solved the problem was writing
something to io_out instead of only io_err on error. Makes sense
but took a while for me to think to try that. Will check this
other stuff in anyways because it might be more robust even though
it didn't solve the problem.